before changing input system

+++ Assets/Scripts/Combat/CombatRanges.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace OmniGlyph.Combat {
-    public enum CombatRanges : short {
+    public enum CombatRanges : byte {
         Close = 1,
         Mid = 2,
         Far = 4
