+++ Assets/Scripts/Combat/Side.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace OmniGlyph.Combat {
-    public enum Side : short {
+    public enum Side : byte {
         Left = 1,
         Right = 2,
         Middle = 3,
